fix: :twisted_rightwards_arrows: Fix merge issues
Removed db:create; excanged with docker compose up. Updated seeding files to pass new linting rules and use correct imports.
