+++ .github/workflows/pull-request-check.yml
@@ -43,8 +43,6 @@ jobs:
         run: docker compose up -d
       - name: Generate migration files
         run: pnpm db:generate
-      - name: Create database
-        run: pnpm db:create
       - name: Migrate database
         run: pnpm db:migrate
       - name: Build
@@ -77,8 +75,6 @@ jobs:
         run: docker compose up -d
       - name: Generate migration files
         run: pnpm db:generate
-      - name: Create database
-        run: pnpm db:create
       - name: Migrate database
         run: pnpm db:migrate
       - name: Seed database
